Add api_key header for temporal requests and fix domain delete api

+++ backend/backend/app/controllers/auth_router.py
@@ -1,5 +1,6 @@
 """Auth controller implementation."""
 import logging
+from http import HTTPStatus
 from typing import Optional
 
 from classy_fastapi import Routable, get, post, delete
@@ -8,6 +9,7 @@ from starlette.requests import Request
 from starlette.responses import RedirectResponse, Response
 
 from backend.app.container import container
+from backend.app.exceptions import HTTPException
 from backend.app.models.dtos.user_status_dto import UserStatusDto
 from backend.app.router import router
 from backend.app.services.auth_cookie_service import (
@@ -22,6 +24,7 @@ from backend.app.services.user_service import (
     get_access_token,
     get_refresh_token,
 )
+from backend.config import settings
 from common.models.user import User, UserLoginWithOTP
 
 log = logging.getLogger(__name__)
@@ -124,6 +127,11 @@ class AuthRoutes(Routable):
         response: Response,
         user: User = Depends(get_logged_user),
     ):
+        if request.headers.get("api_key") != settings.temporal_settings.api_key:
+            raise HTTPException(
+                status_code=HTTPStatus.FORBIDDEN,
+                content="You are not allowed to perform this action.",
+            )
         await self.auth_service.delete_user(user=user)
         delete_token_cookie(response=response)
         await add_refresh_token_to_blacklist(request=request)
@@ -136,6 +144,6 @@ class AuthRoutes(Routable):
         refresh_token=Depends(get_refresh_token),
         user: User = Depends(get_logged_user),
     ):
-        await self.auth_service.add_workflow_to_delete_user(
+        return await self.auth_service.add_workflow_to_delete_user(
             access_token=access_token, refresh_token=refresh_token, user=user
         )
